refactor(frontend): split TodoPostPage into sub-components
Split TodoPostPage (925→1 line) into:
- todo-post/CollapsibleCommand.tsx: collapsible command display
- todo-post/RatingControl.tsx: execution rating control
- todo-post/WorktreePathDisplay.tsx: worktree path display
- todo-post/ReplyRow.tsx: reply input row
- todo-post/PostCard.tsx: main post card
- todo-post/ThreadGroup.tsx: thread group component
- todo-post/LogDrawer.tsx: log/chat/command drawer
- todo-post/helpers.ts: re-exports from todo-detail/helpers
- todo-post/index.tsx: main panel

All exports maintained via re-export from original file.
