+++ frontend/src/components/todo-post/helpers.ts
@@ -1,0 +1,5 @@
+// todo-post 组件的辅助函数。
+
+// 从 todo-detail/helpers 重新导出所需的工具函数
+export { getElapsedSeconds, groupBySession, formatLogTime, hasLogsStatic } from '../todo-detail/helpers';
+export type { SessionGroup } from '../todo-detail/helpers';
